+++ pioreactor/cli/pio.py
@@ -263,12 +263,12 @@ def version(verbose: bool) -> None:
         from pioreactor.version import tuple_to_text
         from pioreactor.version import get_firmware_version
         from pioreactor.version import rpi_version_info
-        from pioreactor.version import _get_product_from_id
+        from pioreactor.version import get_product_from_id
 
         click.echo(f"Pioreactor software:    {tuple_to_text(software_version_info)}")
         click.echo(f"Pioreactor HAT:         {tuple_to_text(hardware_version_info)}")
         click.echo(f"Pioreactor firmware:    {tuple_to_text(get_firmware_version())}")
-        click.echo(f"Model name:             {_get_product_from_id()}")
+        click.echo(f"Model name:             {get_product_from_id()}")
         click.echo(f"HAT serial number:      {serial_number}")
         click.echo(f"Operating system:       {platform.platform()}")
         click.echo(f"Raspberry Pi:           {rpi_version_info}")